+++ Codebase/equaly/src/main/java/org/hwr/equaly/controller/setFormatter/SetFormatterImpl.java
@@ -1,12 +1,22 @@
-package org.hwr.equaly.controller.textSplitter;
+package org.hwr.equaly.controller.setFormatter;
 
 import org.hwr.equaly.model.Fragment;
 import org.springframework.stereotype.Component;
 import java.util.ArrayList;
 
+/**
+ * Implementing logic solely to safely merge together a token set its associated pos-tag set
+ */
 @Component
-public class TextSplitterImpl implements TextSplitter {
+public class SetFormatterImpl implements SetFormatter {
 
+    /**
+     * Given both a set of tokens and an (ideally) equally sized set of associated taggings, safely merge both these
+     * sets together into one 2D structure of Fragment objects for further processing
+     * @param tokens these were made from the input text
+     * @param tags these were made from the input text's tokens
+     * @return a unified data structure now enabling replacement
+     */
     @Override
     public Fragment[][] createSubsets(String[] tokens, String[] tags) {
         ArrayList<Fragment[]> sentences = new ArrayList<>();
